pyproject.toml
[project]
name = "transcribe-diarize-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "deepmultilingualpunctuation>=1.0.1",
    "fastapi>=0.117.1",
    "faster-whisper>=1.2.0",
    "huggingface>=0.0.1",
    "nemo-toolkit>=2.4.0",
    "numpy>=2.3.3",
    "pyannote-audio>=3.4.0",
    "pydub>=0.25.1",
    "ruff>=0.13.1",
    "scipy>=1.16.2",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "transformers>=4.56.2",
    "uvicorn>=0.36.0",
    "wget>=3.2",
    "whisperx>=3.4.2",
]
